refactor: 로그인 이전 화면에서 Gradient Background를 분리 후 layout에서 적용하도록 리팩토링
