+++ app/(beforeLogin)/login/page.tsx
@@ -1,30 +1,17 @@
 "use client"
 import Image from "next/image";
-import { FormEvent, useEffect, useState } from "react";
+import { FormEvent, useState } from "react";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import { useRouter } from "next/navigation";
 import { LoginRequest } from "@/dto/loginDTO";
 import { postLogin } from "@/services/api/postLogin";
-import { Background1, Background2, Logo } from "@/assets/images";
-
-const backgroundImages = [
-  Background1, Background2
-]
+import { Logo } from "@/assets/images";
 
 export default function LoginPage() {
   const router = useRouter();
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
-  const [index, setIndex] = useState(0);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setIndex((prevIndex) => (prevIndex + 1) % backgroundImages.length);
-    }, 3000);
-
-    return () => clearInterval(interval);
-  }, []);
 
   const handleSubmit = async (e: FormEvent) => {
     e.preventDefault();
@@ -51,21 +38,6 @@ export default function LoginPage() {
 
   return (
     <div className="relative min-h-screen flex justify-center items-center overflow-hidden">
-      {/* 배경 이미지 컨테이너 */}
-      <div className="absolute inset-0 w-full h-full z-0">
-        {backgroundImages.map((src, i) => (
-          <Image
-            key={i}
-            src={src}
-            alt={`background-${i}`}
-            layout="fill"
-            objectFit="cover"
-            className={`absolute inset-0 transition-opacity duration-1000 ${i === index ? "opacity-100" : "opacity-0"
-              }`}
-          />
-        ))}
-      </div>
-
       <div className="z-10 w-[420px] h-[520px] flex flex-col gap-16 justify-center items-center m-auto py-32 rounded-3xl shadow-lg bg-white bg-opacity-60">
         <Image src={Logo} alt="logo" width={300} height={200} />
         <form className="flex flex-col justify-center w-80" onSubmit={handleSubmit}>
